app/requirements.txt
flask
uwsgi
mysql-connector-python